+++ ecs.php
@@ -7,7 +7,7 @@ return static function(ECSConfig $ecsConfig): void {
     $ecsConfig->paths([
         __DIR__ . '/examples',
         __DIR__ . '/src',
-        __DIR__ . '/tests/pest',
+        __DIR__ . '/tests',
         __FILE__,
     ]);
     $ecsConfig->parallel();